squashed commits from 18313
