+++ doc/src/modules/matrices/standalone.rst
@@ -1,0 +1,46 @@
+Matrix Standalone Functions
+===========================
+
+The following top-level standalone functions are available for importing from
+``sympy.matrices``. They replicate the functionality of matrix methods of the
+same name but with the form ``func(M)`` instead of ``M.func()``.
+
+adjugate
+--------
+.. autofunction:: sympy.matrices.adjugate
+
+charpoly
+--------
+.. autofunction:: sympy.matrices.charpoly
+
+cofactor
+--------
+.. autofunction:: sympy.matrices.cofactor
+
+cofactor_matrix
+---------------
+.. autofunction:: sympy.matrices.cofactor_matrix
+
+det
+---
+.. autofunction:: sympy.matrices.det
+
+det_bareiss
+-----------
+.. autofunction:: sympy.matrices.det_bareiss
+
+det_berkowitz
+-------------
+.. autofunction:: sympy.matrices.det_berkowitz
+
+det_LU
+------
+.. autofunction:: sympy.matrices.det_LU
+
+minor
+-----
+.. autofunction:: sympy.matrices.minor
+
+minor_submatrix
+---------------
+.. autofunction:: sympy.matrices.minor_submatrix
